refactor: updated layout structure and landing steps; adjusted Tailwind container config
