+++ src/app/layout.tsx
@@ -1,7 +1,9 @@
 import type { Metadata } from "next";
 import { Inter } from "next/font/google";
+import { Logirent } from "./fonts/logirent";
 import "./globals.css";
 import { cn } from "@/lib/utils";
+import Navbar from "@/components/ui/Navbar";
 
 const inter = Inter({
   variable: "--font-inter-sans",
@@ -20,13 +22,9 @@ export default function RootLayout({
   children: React.ReactNode;
 }>) {
   return (
-    <html lang="en" className="light">
-      <body
-        className={cn(
-          "min-h-screen font-sans antialiased grainy",
-          inter.className
-        )}
-      >
+    <html lang="en" className={cn(inter.variable, Logirent.variable, "light")}>
+      <body className={cn("min-h-screen font-sans antialiased grainy")}>
+        <Navbar />
         {children}
       </body>
     </html>
