+++ src/app/page.tsx
@@ -1,114 +1,9 @@
-// import MaxWidthWrapper from "@/components/MaxWidthWrapper";
-// import { buttonVariants } from "@/components/ui/button";
-// import { cn } from "@/lib/utils";
-// import { ArrowRight } from "lucide-react";
-// import Image from "next/image";
-// import Link from "next/link";
-
-// export default function Home() {
-//   return (
-//     <>
-//       <MaxWidthWrapper className="mb-2 mt-28 sm:mt-40 flex flex-col items-center justify-center">
-//         <div className="mx-auto mb-4 flex max-w-fit items-center justify-center space-x-2 overflow-hidden rounded-full border border-gray-200 bg-white px-7 py-2 shadow-md backdrop-blur transition-all hover:border-gray-300 hover:bg-white/50">
-//           <p className="text-sm font-semibold text-gray-700">
-//             Zeph is now public!
-//           </p>
-//         </div>
-//         <div className="flex flex-col items-center">
-//           <h1 className="max-w-4xl text-5xl text-center  font-bold md:text-6xl lg:text-7xl">
-//             Chat with your <span className="text-blue-600">documents</span> in
-//             seconds.
-//           </h1>
-
-//           <p className="mt-5 max-w-prose text-center text-zinc-700 sm:text-lg">
-//             Zeph allows you to have conversations with any PDF document. Simply
-//             upload your file and start asking questionns right away.
-//           </p>
-
-//           <Link
-//             href="/dashboard"
-//             target="_blank"
-//             className={cn(
-//               "mt-6 flex text-center",
-//               buttonVariants({
-//                 size: "lg",
-//               })
-//             )}
-//           >
-//             <p>Get Started</p>
-//             <ArrowRight className="h-5 w-5" />
-//           </Link>
-//         </div>
-//       </MaxWidthWrapper>
-
-//       {/* Value Proposition section */}
-//       <div>
-//         <div className="relative isolate">
-//           <div
-//             aria-hidden="true"
-//             className="pointer-events-none absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80"
-//           >
-//             <div
-//               style={{
-//                 clipPath:
-//                   "polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)",
-//               }}
-//               className="relative left-[calc(50%-11rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30 sm:left-[calc(50%-30rem)] sm:w-[72.1875rem]"
-//             />
-//           </div>
-
-//           <div>
-//             <div className="mx-auto max-w-6xl px-6 lg:px-8">
-//               <div className="mt-26 flow-root sm:mt-24">
-//                 <div className="-mt-2 rounded-xl bg-gray-900/5 p-2 ring-1 ring-inset ring-gray-900/2 lg:m-4 lg:rounded-2xl lg:p-4">
-//                   <Image
-//                     src="/dashboard-preview.jpg"
-//                     alt="dashboard image"
-//                     width={1364}
-//                     height={866}
-//                     className="rounded-md bg-white p-2 sm:p-8 md:p-20 shadow-2xl ring-1 ring-gray-900/5"
-//                   />
-//                 </div>
-//               </div>
-//               <div className="mt-26 flow-root sm:mt-24">
-//                 <div className="-mt-2 rounded-xl bg-gray-900/5 p-2 ring-1 ring-inset ring-gray-900/2 lg:m-4 lg:rounded-2xl lg:p-4">
-//                   <Image
-//                     src="/file-upload-preview.jpg"
-//                     alt="dashboard image"
-//                     width={1364}
-//                     height={866}
-//                     className="rounded-md bg-white p-2 sm:p-8 md:p-20 shadow-2xl ring-1 ring-gray-900/5"
-//                   />
-//                 </div>
-//               </div>
-//             </div>
-//           </div>
-
-//           <div
-//             aria-hidden="true"
-//             className="pointer-events-none absolute inset-x-0 -top-40 -z-10 transform-gpu overflow-hidden blur-3xl sm:-top-80"
-//           >
-//             <div
-//               style={{
-//                 clipPath:
-//                   "polygon(74.1% 44.1%, 100% 61.6%, 97.5% 26.9%, 85.5% 0.1%, 80.7% 2%, 72.5% 32.5%, 60.2% 62.4%, 52.4% 68.1%, 47.5% 58.3%, 45.2% 34.5%, 27.5% 76.7%, 0.1% 64.9%, 17.9% 100%, 27.6% 76.8%, 76.1% 97.7%, 74.1% 44.1%)",
-//               }}
-//               className="relative left-[calc(50%-13rem)] aspect-[1155/678] w-[36.125rem] -translate-x-1/2 rotate-[30deg] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30 sm:left-[calc(50%-36rem)] sm:w-[72.1875rem]"
-//             />
-//           </div>
-//         </div>
-//       </div>
-//     </>
-//   );
-// }
-
-import MaxWidthWrapper from "@/components/MaxWidthWrapper";
+import MaxWidthWrapper from "@/components/ui/MaxWidthWrapper";
 import Link from "next/link";
 import { ArrowRight } from "lucide-react";
-import { buttonVariants } from "@/components/ui/button";
+import { buttonVariants } from "@/components/ui/Button";
 import Image from "next/image";
-import Step from "@/components/Step";
-import Steps from "@/components/Steps";
+import Steps from "@/components/landing/Steps";
 
 export default function Home() {
   return (
