+++ src/components/landing/Steps.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import Step from "@/components/Step";
+import Step from "@/components/landing/Step";
 
 const Steps = () => {
   return (
